feat: Step3 선택사항입력 화면 구현
