+++ frontend/app/src/main/java/com/apptive/japkor/ui/requiredinfo/steps/Step2Content.kt
@@ -1,11 +1,7 @@
 package com.apptive.japkor.ui.requiredinfo.steps
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -14,7 +10,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -22,13 +17,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.apptive.japkor.ui.components.CustomOutlinedTextField
 import com.apptive.japkor.ui.components.CustomText
 import com.apptive.japkor.ui.components.CustomTextType
 import com.apptive.japkor.ui.theme.CustomColor
+import com.apptive.japkor.ui.components.OptionChip
 
 @Composable
 fun Step2Content() {
@@ -211,50 +206,5 @@ fun Step2Content() {
             modifier = Modifier.padding(horizontal = 7.dp),
             size = 14.sp
         )
-    }
-}
-
-@Composable
-fun OptionChip(
-    text: String,
-    selected: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val shape = RoundedCornerShape(12.dp)
-
-    val background = if (selected) CustomColor.gray400 else CustomColor.white
-    val textColor = if (selected) CustomColor.white else CustomColor.gray300
-    val borderColor = if (selected) CustomColor.gray300 else CustomColor.gray200
-
-    Surface(
-        modifier = modifier
-            .height(44.dp)
-            .border(
-                width = 1.dp,
-                color = borderColor,
-                shape = shape
-            )
-            .clip(shape)
-            .clickable(
-                indication = null,
-                interactionSource = remember { MutableInteractionSource() },
-                onClick = onClick
-            ),
-        color = background,
-        shape = shape,
-        shadowElevation = 0.dp
-    ) {
-        Box(
-            modifier = Modifier.padding(horizontal = 12.dp),
-            contentAlignment = Alignment.Center
-        ) {
-            CustomText(
-                text = text,
-                type = CustomTextType.body,
-                size = 14.sp,
-                color = textColor
-            )
-        }
     }
 }
